+++ CHANGELOG.md
@@ -10,9 +10,10 @@
 
 ### Bug Fixes
 
-* **card:** move hide snoozed toggle onto the filter row ([#499](https://github.com/mossipcams/autosnooze/issues/499)) ([4474837](https://github.com/mossipcams/autosnooze/commit/4474837b43f29810d54f02f4bfa356cbd881fffe))
+* **card:** move hide snoozed toggle onto the header row ([#508](https://github.com/mossipcams/autosnooze/issues/508))
 * import StaticPathConfig on HA without http.server ([#505](https://github.com/mossipcams/autosnooze/issues/505)) ([9350c25](https://github.com/mossipcams/autosnooze/commit/9350c25294ae0b35dcef3dfe43655ba78427f135))
 * **telemetry:** harden PostHog privacy and platform reporting ([#507](https://github.com/mossipcams/autosnooze/issues/507)) ([0609fe3](https://github.com/mossipcams/autosnooze/commit/0609fe396e612b40e224c90bda1fb16a0da2744a))
+* **telemetry:** use the correct PostHog project API key ([#508](https://github.com/mossipcams/autosnooze/issues/508))
 
 ## [0.2.29](https://github.com/mossipcams/autosnooze/compare/v0.2.28...v0.2.29) (2026-08-11)